remove sidebar from project page
