+++ styles/ProjectHead.module.css
@@ -1,0 +1,38 @@
+.headHolder {
+  overflow: hidden;
+  position: relative;
+}
+
+.backgroundCard {
+  filter: sepia(30%) blur(8px);
+  width: 100%;
+  height: 250px;
+  object-fit: cover;
+}
+
+.headInfo h1 {
+  position: absolute;
+  left: 0;
+  bottom: 0;
+
+  margin: 0;
+  padding: 0 4px;
+
+  background-color: rgba(255, 255, 255, .8);
+}
+
+.links {
+  position: absolute;
+  top: 0;
+  right: 0;
+  background-color: white;
+}
+
+.links a {
+  font-size: 20px;
+  padding: 4px 4px;
+}
+
+.links a > * {
+  color: black;
+}
