--- styles/ProjectInfo.module.css
@@ -1,95 +1,0 @@
-.project {
-  background-color: #fff;
-}
-
-.articles > * {
-  padding: 16px 12px 0;
-}
-
-.members {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(150px, auto));
-}
-
-.members img {
-  width: 80px;
-  height: 80px;
-  object-fit: cover;
-  margin-bottom: 5px;
-}
-
-.memberCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 155px;
-}
-
-.techs {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(80px, auto));
-}
-
-.techs img {
-  width: 60px;
-  height: 60px;
-  object-fit: contain;
-  background-color: #ffffff;
-}
-
-.techCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 140px;
-}
-
-.techs a, .techs a:hover {
-  color: #000000;
-}
-
-#year {
-  color: #444;
-  float: right;
-}
-
-/* heading */
-
-.headHolder {
-  overflow: hidden;
-  position: relative;
-}
-
-.backgroundCard {
-  filter: sepia(30%) blur(8px);
-  width: 100%;
-  height: 250px;
-  object-fit: cover;
-}
-
-.headInfo h1 {
-  position: absolute;
-  left: 0;
-  bottom: 0;
-
-  margin: 0;
-  padding: 0 4px;
-
-  background-color: rgba(255, 255, 255, .8);
-}
-
-.links {
-  position: absolute;
-  top: 0;
-  right: 0;
-  background-color: white;
-}
-
-.links a {
-  font-size: 20px;
-  padding: 4px 4px;
-}
-
-.links a > * {
-  color: black;
-}
