+++ styles/Projects.module.css
@@ -1,9 +1,57 @@
-.projectWrapper {
-  display: grid;
-  grid-template-columns: 28% 72%;
+.project {
+  background-color: #fff;
+  margin: 50px 0 20px;
+  border: 1px solid rgba(200, 200, 200, .25);
+  border-radius: 5px;
 }
 
-.projectOverview {
-  background-color: white;
+.articles > * {
+  padding: 16px 12px 0;
+}
 
+.members {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(150px, auto));
+}
+
+.members img {
+  width: 80px;
+  height: 80px;
+  object-fit: cover;
+  margin-bottom: 5px;
+}
+
+.memberCard {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 155px;
+}
+
+.techs {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(80px, auto));
+}
+
+.techs img {
+  width: 60px;
+  height: 60px;
+  object-fit: contain;
+  background-color: #ffffff;
+}
+
+.techCard {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 140px;
+}
+
+.techs a, .techs a:hover {
+  color: #000000;
+}
+
+#year {
+  color: #444;
+  float: right;
 }
